fix: separate ibid tables
